:recycle: refactored code
