+++ src/main/java/by/petrovich/eshop/EShopApplication.java
@@ -1,9 +1,7 @@
 package by.petrovich.eshop;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
@@ -12,10 +10,4 @@ public class EShopApplication {
     public static void main(String[] args) {
         SpringApplication.run(EShopApplication.class, args);
     }
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-
 }
