+++ src/main/java/by/petrovich/eshop/config/WebSecurityConfig.java
@@ -1,7 +1,6 @@
 package by.petrovich.eshop.config;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -13,7 +12,6 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
 @EnableWebSecurity
-@ComponentScan(basePackages = "by")
 public class WebSecurityConfig {
 
     @Bean
