+++ src/main/java/by/petrovich/eshop/controller/GoToController.java
@@ -1,45 +1,26 @@
-package by.petrovich.eshop.controllers;
+package by.petrovich.eshop.controller;
 
-import by.petrovich.eshop.dto.CartDto;
-import by.petrovich.eshop.entity.User;
 import by.petrovich.eshop.service.CategoryService;
 import by.petrovich.eshop.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
-import static by.petrovich.eshop.PageName.ADMIN_PAGE;
-import static by.petrovich.eshop.PageName.CART_PAGE;
-import static by.petrovich.eshop.PageName.HOME_PAGE;
-import static by.petrovich.eshop.PageName.LOGIN_PAGE;
-import static by.petrovich.eshop.PageName.PROFILE_PAGE;
-import static by.petrovich.eshop.PageName.REGISTRATION_PAGE;
+import static by.petrovich.eshop.utils.PageName.ADMIN_PAGE;
+import static by.petrovich.eshop.utils.PageName.CART_PAGE;
+import static by.petrovich.eshop.utils.PageName.HOME_PAGE;
+import static by.petrovich.eshop.utils.PageName.LOGIN_PAGE;
+import static by.petrovich.eshop.utils.PageName.PROFILE_PAGE;
+import static by.petrovich.eshop.utils.PageName.REGISTRATION_PAGE;
 
-@SessionAttributes({"user", "cartDto"})
 @RestController
+@RequiredArgsConstructor
 public class GoToController {
     private final CategoryService categoryService;
     private final UserService userService;
-
-    @Autowired
-    public GoToController(CategoryService categoryService, UserService userService) {
-        this.categoryService = categoryService;
-        this.userService = userService;
-    }
-
-    @ModelAttribute("user")
-    public User initializeUserSessionObject() {
-        return new User();
-    }
-
-    @ModelAttribute("cartDto")
-    public CartDto initializeCartSessionObject() {
-        return new CartDto();
-    }
 
     @GetMapping("/home")
     public ModelAndView showHomePage() {
