+++ src/main/java/by/petrovich/eshop/controller/UserController.java
@@ -1,11 +1,10 @@
-package by.petrovich.eshop.controllers;
+package by.petrovich.eshop.controller;
 
 import by.petrovich.eshop.dto.CartDto;
 import by.petrovich.eshop.dto.RegistrationFormDto;
 import by.petrovich.eshop.entity.User;
 import by.petrovich.eshop.service.UserService;
 import jakarta.validation.Valid;
-import jakarta.validation.ValidationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -15,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
-import static by.petrovich.eshop.PageName.LOGIN_PAGE;
+import static by.petrovich.eshop.utils.PageName.LOGIN_PAGE;
 
 @Controller
 @SessionAttributes({"user", "cartDto"})
@@ -42,10 +41,6 @@ public class UserController {
     public ModelAndView registrate(@ModelAttribute("registration_form_dto") @Valid RegistrationFormDto registrationFormDto,
                                    BindingResult errors,
                                    ModelAndView model) {
-        if (errors.hasErrors()) {
-            throw new ValidationException((Throwable) errors);
-            // TODO: 3 июл. 2023 г.  
-        }
         model.addObject("registration_form_dto", registrationFormDto);
         userService.register(registrationFormDto);
         return new ModelAndView(LOGIN_PAGE);
