+++ src/main/java/by/petrovich/eshop/exceptions/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.ModelAndView;
 
-import static by.petrovich.eshop.PageName.ERROR_PAGE;
+import static by.petrovich.eshop.utils.PageName.ERROR_PAGE;
 
 @Slf4j
 @ControllerAdvice
@@ -14,8 +14,8 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(UserNotFoundException.class)
     public ModelAndView handleUserNotFoundException(UserNotFoundException ex) {
         ModelAndView modelAndView = new ModelAndView(ERROR_PAGE);
-        modelAndView.addObject("userNotFoundException", ex.getMessage());
-        log.error("error", ex);
+        modelAndView.addObject("User not found", ex.getMessage());
+        log.error("User not found", ex);
         return modelAndView;
     }
