+++ src/main/resources/templates/home.html
@@ -9,7 +9,7 @@
           href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css">
 </head>
 <body>
-<div th:insert="~{include/topnav :: navigation}"></div>
+<div th:insert="~{fragment/topnav :: navigation}"></div>
 <div class="main-content" style="padding-left:80px">
     <h2>Categories</h2>
     <div th:each="c : ${categories}">
@@ -24,5 +24,5 @@
     </div>
 </div>
 </body>
-<div th:insert="~{include/footer :: footer}"></div>
+<div th:insert="~{fragment/footer :: footer}"></div>
 </html>
