+++ src/main/resources/templates/order.html
@@ -10,7 +10,7 @@
           href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css">
 </head>
 <body>
-<div th:insert="~{include/topnav :: navigation}"></div>
+<div th:insert="~{fragment/topnav :: navigation}"></div>
 <div class="order-content">
     <h1 th:text="${order.orderId}">Order</h1>
     <h3 th:text="${order.createdAt}">createdAt</h3>
@@ -35,5 +35,5 @@
     </div>
 </div>
 </body>
-<div th:insert="~{include/footer :: footer}"></div>
+<div th:insert="~{fragment/footer :: footer}"></div>
 </html>
